Configuration

RenoiseOSC defaults to sending events to localhost at port 8000, the default port which Renoise hosts its OSC server on. To use a different address, use the functions below. Only communication over UDP is supported.

RenoiseOSC.setaddress! — Function

setaddress!(host::Union{AbstractString,Sockets.IPAddr}, port::Integer)

Set the host and the port of the OSC Server.

RenoiseOSC.sethost! — Function

sethost!(host::Union{AbstractString,Sockets.IPAddr})

Set the host of the Renoise OSC server. Default value is ip"127.0.0.1".

Julia 1.3

Setting the host by passing an AbstractString requires at least Julia 1.3.

Example

julia> sethost!(ip"127.0.0.1")
Sockets.InetAddr{Sockets.IPv4}(ip"127.0.0.1", 8000)

RenoiseOSC.setport! — Function

setport!(port::Integer)

Set the port of the Renoise OSC server. Default value is 8000.