Renoise OSC API

This file documents the API exposed in Renoise ^3.0.1.

v3.0

RenoiseOSC.luaeval — Function

luaeval(expr::AbstractString)

Evaluate a Lua expression inside Renoise's scripting environment.

Example

luaeval("renoise.song().transport.bpm = 132")

RenoiseOSC.tempo — Function

tempo(bpm::Integer)

Set the song's current bpm value. The valid tempo range is 32:999 for <v3.4, 20:999 for >=v3.4.

RenoiseOSC.editmode — Function

editmode(on::Bool)

Set the song's global edit mode on or off.

RenoiseOSC.octave — Function

octave(oct::Integer)

Sets the song's current octave 0:8.

RenoiseOSC.patternfollow — Function

patternfollow(on::Bool)

Enable or disable the global pattern follow mode.

RenoiseOSC.editstep — Function

editstep(step::Integer)

Set the song's current edit step [0:8].

RenoiseOSC.linesperbeat — Function

linesperbeat(lpb::Integer)

Set the song's current lines per beat 1:255.

RenoiseOSC.metronome — Function

metronome(on::Bool)

Enable or disable the metronome.

RenoiseOSC.metronomeprecount — Function

metronomeprecount(on::Bool)

Enable or disable the global metronome precount.

RenoiseOSC.quantization — Function

quantization(on::Bool)

Enable or disable the global record quantization.

RenoiseOSC.quantizationstep — Function

quantizationstep(step::Integer)

Set the global record quantization step 1:32.

RenoiseOSC.scheduleadd — Function

scheduleadd(sequence::Integer)

Add a scheduled sequence playback position.

RenoiseOSC.scheduleset — Function

scheduleset(sequence::Integer)

Replace the current scheduled sequence playback position.

RenoiseOSC.slotmute — Function

slotmute(track::Integer, sequence::Integer)
slotmute(mute::Bool, track::Integer, sequence::Integer)

Mute the specified track at the specified sequence slot in the matrix.

RenoiseOSC.slotunmute — Function

slotunmute(track::Integer, sequence::Integer)

Unmute the specified track at the specified sequence slot in the matrix.

RenoiseOSC.trigger — Function

trigger(sequence::Integer)

Set playback pos to the specified sequence position.

RenoiseOSC.bypass — Function

bypass(bypassed::Bool; track::Integer=-1, device::Integer=-1)

Enable or disable the bypassing of an effect device, set true to bypass the device. When unspecified, track and device default to the currently selected track and device.

RenoiseOSC.setparam — Function

setparam(key::Union{AbstractString,Integer}, value::Real; track::Integer=-1, device::Integer=-1)

Set the parameter of any effect device by its name or index. value is clamped between 0.0 and 1.0. When unspecified, track and device default to the currently selected track and device.

RenoiseOSC.outputdelay — Function

outputdelay(δt::Real; track::Integer)

Set the given track's pre-FX output delay in milliseconds [-100:100]. Default to the currently selected track.

RenoiseOSC.postfxpanning — Function

postfxpanning(pan::Integer; track::Integer=-1)

Set track's post-FX panning, -50:50 left to right. Default to the currently selected track.

RenoiseOSC.postfxvolume — Function

postfxvolume(level::Real; track::Integer=-1)

Set tracks post-FX volume, 0.0:db2lin(6.0), -Inf:+6dB. Default to the currently selected track.

RenoiseOSC.postfxvolumedb — Function

postfxvolumedb(level::Real; track::Integer=-1)

Set tracks post-FX volume in decibels, -200.0:3.0. Default to the currently selected track.

RenoiseOSC.prefxpanning — Function

prefxpanning(pan::Integer; track::Integer=-1)

Set track's pre-FX panning, [-50:50] left to right. Default to the currently selected track.

RenoiseOSC.prefxvolume — Function

prefxvolume(level::Real; track::Integer=-1)

Set tracks pre-FX volume, [0.0:db2lin(6.0)], -Inf:+6dB. Default to the currently selected track.

RenoiseOSC.prefxvolumedb — Function

prefxvolumedb(level::Real; track::Integer=-1)

Set tracks pre-FX volume in decibels, -200.0:3.0. Default to the currently selected track.

RenoiseOSC.prefxwidth — Function

prefxwidth(width::Real; track::Integer=-1)

Set track's pre-FX width 0.0:1.0. Default to the currently selected track.

RenoiseOSC.solo — Function

solo(; track::Integer)

Toggle solo for the given track. Default to the currently selected track.

RenoiseOSC.mute — Function

mute(muted::Bool = true; track::Integer = -1)

Mute the given track. Default to the currently selected track.

RenoiseOSC.unmute — Function

unmute(; track::Integer=-1)
mute(false; track::Integer=-1)

Unmute the given track. Default to the currently selected track.

RenoiseOSC.panic — Function

panic()

Stop playback and reset all playing instruments and DSPs

RenoiseOSC.resume — Function

resume()

Continue playback. Called continue in the Renoise docs, but this name can't be used as continue is a reserved word in Julia.

RenoiseOSC.start — Function

start()

Start song playback or restart playing the current pattern.

RenoiseOSC.stop — Function

stop()

Stop song playback.

RenoiseOSC.midievent — Function

midievent(portid::UInt8, status::UInt8, data1::UInt8, data2::UInt8)
midievent(event::Union{UInt32,UInt64,Array{<:Integer},NTuple{4,<:Integer}})

Fire a raw MIDI event.

RenoiseOSC.noteon — Function

noteon(pitch::Integer, velocity::Integer; instrument::Integer=-1, track::Integer=-1)

Turn on the note with the velocity. Default to the currently selected instrument and track.

RenoiseOSC.noteoff — Function

noteoff(pitch::Integer, velocity::Integer; instrument::Integer=-1, track::Integer=-1)

Trigger a note off.

v3.3

Below are the functions added between Renoise 3.0.1 and 3.3.0.

RenoiseOSC.setmacroparam — Function

setmacroparam(param::Integer, value::Real; instrument::Integer = -1)

Set instrument's macro parameter value 0.0:1.0. Defaults to the currently selected instrument and macro 1.

RenoiseOSC.monophonic — Function

monophonic(mono::Bool; instrument::Integer=-1)

Enable or disable instrument's mono mode. Default to the currently selected instrument.

RenoiseOSC.monophonicglide — Function

monophonicglide(glide::Integer; instrument::Integer=-1)

Set instrument's glide amount 0:255. Default to the currently selected instrument.

RenoiseOSC.phraseplayback — Function

phraseplayback(mode::AbstractString; instrument::Integer=-1)

Set instrument's phrase playback mode ["Off", "Program", "Keymap"] Default to the currently selected instrument.

RenoiseOSC.phraseprogram — Function

phraseprogram(program::Integer; instrument::Integer=-1)

Set instrument's phrase program number 0:127. Default to the currently selected instrument.

RenoiseOSC.quantizationmode — Function

quantizationmode(mode::AbstractString; instrument::Integer=-1)

Set instrument's quantization method, one of "None", "Line", "Beat", "Bar". Default to the currently selected instrument.

RenoiseOSC.scalekey — Function

scalekey(key::AbstractString; instrument::Integer=-1)

Set instrument's note scaling key ["C", "C#"..., "B"]. Default to the currently selected instrument.

RenoiseOSC.scalemode — Function

scalemode(mode::AbstractString; instrument::Integer=-1)

Set instrument's note scaling mode. Default to the currently selected instrument.

RenoiseOSC.transpose — Function

transpose(pitch::Integer; instrument::Integer=-1)

Set instrument's global pitch transpose -120:120. Default to the currently selected instrument.

RenoiseOSC.volume — Function

volume(level::Real; instrument::Integer=-1)

Set instrument's global volume to level [0.0:db2lin(6)]≈[0.0:2.0]. Default to the currently selected instrument.

RenoiseOSC.volumedb — Function

volumedb(level::Real; instrument::Integer=-1)

Set instrument's global volume to level in decibels [0.0:6.0]. Default to the currently selected instrument.

RenoiseOSC.ticksperline — Function

ticksperline(tpl::Integer)

Set the song's current ticks per line 1:16.

RenoiseOSC.loopblock — Function

loopblock(loop::Bool)

Enable or disable pattern block loop.

RenoiseOSC.loopblockmovebackwards — Function

loopblockmovebackwards()

Move the block loop one segment backwards.

RenoiseOSC.loopblockmoveforwards — Function

loopblockmoveforwards()

Move the block loop one segment forwards.

RenoiseOSC.looppattern — Function

looppattern(loop::Bool)

Enable or disable looping the current pattern.

v3.4

Below are functions added in v3.4.

loopsequence